Move reprocess_archive_stubs task to form_processor/tasks.py
